+++ src/marais/graphql/dsl/fetchers.kt
@@ -7,8 +7,10 @@ import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.future.asCompletableFuture
+import kotlinx.coroutines.future.await
 import kotlinx.coroutines.future.future
 import kotlinx.coroutines.reactive.asPublisher
+import java.util.concurrent.CompletableFuture
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.reflect.KFunction
@@ -28,7 +30,7 @@ fun KFunction<*>.fetcher(
     schemaCtx: SchemaBuilderContext,
     scope: CoroutineScope = GlobalScope,
     corCtx: CoroutineContext = EmptyCoroutineContext,
-): DataFetcher<Any> {
+): DataFetcher<Any?> {
     return if (this.isSuspend) when (returnType.classifier) {
         Map::class -> DataFetcher { env ->
             val args = args.resolve(env)
@@ -64,9 +66,10 @@ fun KFunction<*>.fetcher(
                 DataFetcher { env ->
                     val args = args.resolve(env)
                     scope.future(corCtx) {
-                        (this@fetcher as KFunction<Flow<Any>>).callSuspend(env.getSource() ?: receiver, *args)
-                            .map { (it as Map<Any, Any?>).entries }
-                            .asPublisher(coroutineContext)
+                        (this@fetcher as KFunction<Flow<Map<Any, Any?>>>).callSuspend(
+                            env.getSource() ?: receiver,
+                            *args
+                        ).map { it.entries }.asPublisher(coroutineContext)
                     }
                 }
             else
@@ -86,22 +89,31 @@ fun KFunction<*>.fetcher(
                             .map { (it as Map<Any, Any?>).entries }
                     }
                 }
-            else
-                DataFetcher { env ->
-                    val args = args.resolve(env)
-                    scope.future(corCtx) {
-                        (this@fetcher as KFunction<Flow<Any>>).callSuspend(env.getSource() ?: receiver, *args)
-                    }
-                }
+            else defaultFetcherSuspend(this, receiver, args, scope, corCtx)
 
-
-        else ->
+        // CompletableFuture is a special value, we can't return CompletableFuture<CompletableFuture>
+        // Since the behaviour is to convert the suspend function to a future, we need to unwrap it.
+        CompletableFuture::class -> if (returnType.unwrap().classifier == Map::class)
             DataFetcher { env ->
                 val args = args.resolve(env)
                 scope.future(corCtx) {
-                    this@fetcher.callSuspend(env.getSource() ?: receiver, *args)
+                    (this@fetcher as KFunction<CompletableFuture<Map<Any, Any?>>>).callSuspend(
+                        env.getSource() ?: receiver, *args
+                    ).await().entries
                 }
             }
+        else
+            DataFetcher { env ->
+                val args = args.resolve(env)
+                scope.future(corCtx) {
+                    // is there a better way of doing this ?
+                    (this@fetcher as KFunction<CompletableFuture<Any?>>).callSuspend(
+                        env.getSource() ?: receiver, *args
+                    ).await()
+                }
+            }
+
+        else -> defaultFetcherSuspend(this, receiver, args, scope, corCtx)
 
     } else when (returnType.classifier) {
         Map::class ->
@@ -127,8 +139,8 @@ fun KFunction<*>.fetcher(
             if (returnType.unwrap().classifier == Map::class)
                 DataFetcher { env ->
                     val args = args.resolve(env)
-                    (this@fetcher as KFunction<Flow<Any>>).call(env.getSource() ?: receiver, *args)
-                        .map { (it as Map<Any, Any?>).entries }
+                    (this@fetcher as KFunction<Flow<Map<Any, Any?>>>).call(env.getSource() ?: receiver, *args)
+                        .map { it.entries }
                         .asPublisher(corCtx)
                 }
             else
@@ -144,17 +156,35 @@ fun KFunction<*>.fetcher(
                     (this@fetcher as KFunction<Flow<Any>>).call(env.getSource() ?: receiver, *args)
                         .map { (it as Map<Any, Any?>).entries }
                 }
-            else
-                DataFetcher { env ->
-                    val args = args.resolve(env)
-                    (this@fetcher as KFunction<Flow<Any>>).call(env.getSource() ?: receiver, *args)
-                }
+            else defaultFetcher(this, receiver, args)
 
-
-        else ->
+        CompletableFuture::class -> if (returnType.unwrap().classifier == Map::class)
             DataFetcher { env ->
                 val args = args.resolve(env)
-                this.call(env.getSource() ?: receiver, *args)
+                (this@fetcher as KFunction<CompletableFuture<Map<Any, Any?>>>).call(env.getSource() ?: receiver, *args)
+                    .thenApply { it.entries }
             }
+        else defaultFetcher(this, receiver, args)
+
+        else -> defaultFetcher(this, receiver, args)
+    }
+}
+
+private fun defaultFetcher(func: KFunction<*>, receiver: Any?, args: List<Argument>): DataFetcher<Any?> =
+    DataFetcher { env ->
+        val args = args.resolve(env)
+        func.call(env.getSource() ?: receiver, *args)
+    }
+
+private fun defaultFetcherSuspend(
+    func: KFunction<*>,
+    receiver: Any?,
+    args: List<Argument>,
+    scope: CoroutineScope,
+    context: CoroutineContext
+): DataFetcher<Any?> = DataFetcher { env ->
+    val args = args.resolve(env)
+    scope.future(context) {
+        func.callSuspend(env.getSource() ?: receiver, *args)
     }
 }