+++ src/marais/graphql/dsl/fields.kt
@@ -17,7 +17,7 @@ sealed class FieldSpec(val name: String, val description: String?) {
     /**
      * The code behind this field returning a value
      */
-    abstract val dataFetcher: DataFetcher<Any>
+    abstract val dataFetcher: DataFetcher<Any?>
 
     /**
      * The output type as shown in the schema (e.g. no async containers)
@@ -30,7 +30,7 @@ class CustomFieldSpec(
     description: String?,
     override val outputType: KType,
     override val arguments: List<Argument>,
-    override val dataFetcher: DataFetcher<Any>
+    override val dataFetcher: DataFetcher<Any?>
 ) : FieldSpec(name, description)
 
 internal class SuspendLambdaFieldSpec(
@@ -52,18 +52,20 @@ internal class SuspendLambdaFieldSpec(
         }
     }
 
-    override val dataFetcher: DataFetcher<Any> =
-        Lambdas.indirectCallSuspend(arity).fetcher(reflected.returnType, arguments, receiver)
+    override val dataFetcher: DataFetcher<Any?> =
+        Lambdas.indirectCallSuspend(arity).fetcher(reflected.returnType, arguments, receiver, context)
 }
 
 internal class PropertyFieldSpec<R>(
     property: KProperty1<R, Any?>,
     name: String,
     description: String?,
-    receiver: R? = null
+    receiver: R? = null,
+    context: SchemaBuilderContext
 ) : FieldSpec(name, description ?: property.extractDesc()) {
 
-    override val dataFetcher: DataFetcher<Any> = property.getter.fetcher(property.returnType, emptyList(), receiver)
+    override val dataFetcher: DataFetcher<Any?> =
+        property.getter.fetcher(property.returnType, emptyList(), receiver, context)
     override val outputType: KType = property.returnType.unwrapAsyncType()
     override val arguments: List<Argument> = emptyList()
 }
@@ -85,5 +87,5 @@ internal class FunctionFieldSpec<R>(
         }
     }
 
-    override val dataFetcher: DataFetcher<Any> = func.fetcher(func.returnType, arguments, receiver)
+    override val dataFetcher: DataFetcher<Any?> = func.fetcher(func.returnType, arguments, receiver, context)
 }